interfaces and AdvChannel

+++ src/controller/RgbImageFileIO.java
@@ -64,7 +64,7 @@ class RgbImageFileIO implements ImageFileIO {
       int width = sc.nextInt();
       int height = sc.nextInt();
       int maxValue = sc.nextInt();
-      int[][][] imageData = new int[3][height][width];
+      double[][][] imageData = new double[3][height][width];
       for (int i = 0; i < height; i++) {
         for (int j = 0; j < width; j++) {
           int red = sc.nextInt();
@@ -95,7 +95,7 @@ class RgbImageFileIO implements ImageFileIO {
       if (image != null) {
         int width = image.getWidth();
         int height = image.getHeight();
-        int[][][] imageData = new int[3][height][width];
+        double[][][] imageData = new double[3][height][width];
 
         ColorModel colorModel = image.getColorModel();
         if (colorModel.getColorSpace().getType() == ColorSpace.TYPE_RGB) {
@@ -154,7 +154,7 @@ class RgbImageFileIO implements ImageFileIO {
   }
 
   private void savePPM(String filePath, ImageData imageData) throws IOException {
-    int[][][] data = imageData.getData();
+    double[][][] data = imageData.getData();
     int width = data[0][0].length;
     int height = data[0].length;
 
@@ -164,9 +164,9 @@ class RgbImageFileIO implements ImageFileIO {
     writer.write(imageData.getMaxValue() + "\n");
     for (int y = 0; y < height; y++) {
       for (int x = 0; x < width; x++) {
-        int red = data[0][y][x];
-        int green = data[1][y][x];
-        int blue = data[2][y][x];
+        int red = (int) data[0][y][x];
+        int green = (int) data[1][y][x];
+        int blue = (int) data[2][y][x];
         writer.write(red + "\n" + green + "\n" + blue + "\n");
       }
     }
@@ -175,7 +175,7 @@ class RgbImageFileIO implements ImageFileIO {
 
   private void saveGeneralFormat(String imageFormat, String destinationPath, ImageData imageData)
           throws IOException {
-    int[][][] pixelValues = imageData.getData();
+    double[][][] pixelValues = imageData.getData();
     int width = pixelValues[0][0].length;
     int height = pixelValues[0].length;
 
@@ -183,7 +183,7 @@ class RgbImageFileIO implements ImageFileIO {
 
     for (int y = 0; y < height; y++) {
       for (int x = 0; x < width; x++) {
-        int rgb = pixelValues[0][y][x] << 16 | pixelValues[1][y][x] << 8 | pixelValues[2][y][x];
+        int rgb =  (int)pixelValues[0][y][x] << 16 | (int)pixelValues[1][y][x] << 8 | (int)pixelValues[2][y][x];
         image.setRGB(x, y, rgb);
       }
     }
